+++ src/server.ts
@@ -1,0 +1,4 @@
+import { GoalPlugin } from "./index.js";
+
+export { GoalPlugin };
+export default GoalPlugin;